Use Spock @IgnoreIf
