+++ .gitignore
@@ -20,3 +20,5 @@ build/
 *.ipr
 *.iws
 .idea/
+/.DS_Store
+/classes
